+++ modules/Printers/src/asciiprinter.cpp
@@ -446,24 +446,24 @@ is a unique record for every rank/pointID pair.";
    
     // Template for print functions of "easy" types
     template<class T>
-    void asciiPrinter::template_print(T const& value, const std::string& label, const int IDcode, const int thread, const int pointID)
+    void asciiPrinter::template_print(T const& value, const std::string& label, const int IDcode, const uint thread, const ulong pointID)
     {
       std::vector<double> vdvalue(1,value); // For now everything has to end up as a vector of doubles
       std::vector<std::string> labels(1,label);
       addtobuffer(vdvalue,labels,IDcode,thread,pointID);       
     }
 
-    void asciiPrinter::print(int const& value, const std::string& label, const int IDcode, const int thread, const int pointID)
+    void asciiPrinter::print(int const& value, const std::string& label, const int IDcode, const uint thread, const ulong pointID)
     { template_print(value,label,IDcode,thread,pointID); }
-    void asciiPrinter::print(double const& value, const std::string& label, const int IDcode, const int thread, const int pointID)
+    void asciiPrinter::print(double const& value, const std::string& label, const int IDcode, const uint thread, const ulong pointID)
     { template_print(value,label,IDcode,thread,pointID); }
     #ifndef STANDALONE  // Need to disable print functions for these if STANDALONE is defined (see baseprinter.hpp line ~41)
-    void asciiPrinter::print(unsigned int const& value, const std::string& label, const int IDcode, const int thread, const int pointID)
+    void asciiPrinter::print(unsigned int const& value, const std::string& label, const int IDcode, const uint thread, const ulong pointID)
     { template_print(value,label,IDcode,thread,pointID); }
     #endif 
     // etc. as needed... 
 
-    void asciiPrinter::print(std::vector<double> const& value, const std::string& label, const int IDcode, const int thread, const int pointID)
+    void asciiPrinter::print(std::vector<double> const& value, const std::string& label, const int IDcode, const uint thread, const ulong pointID)
     {
       std::vector<std::string> labels;
       labels.reserve(value.size());
@@ -477,7 +477,7 @@ is a unique record for every rank/pointID pair.";
       addtobuffer(value,labels,IDcode,thread,pointID);
     }
    
-    void asciiPrinter::print(ModelParameters const& value, const std::string& label, const int IDcode, const int thread, const int pointID)
+    void asciiPrinter::print(ModelParameters const& value, const std::string& label, const int IDcode, const uint thread, const ulong pointID)
     {
       std::map<std::string, double> parameter_map = value.getValues();
       std::vector<std::string> names;
